+++ src/behavior/collapse-expand.ts
@@ -9,7 +9,7 @@ export default {
        * 发生收缩/扩展变化时的回调
        */
       trigger: DEFAULT_TRIGGER,
-      onChange() {},
+      onChange() { },
     };
   },
   getEvents(): { [key in G6Event]?: string } {
@@ -22,6 +22,7 @@ export default {
       // eslint-disable-next-line no-console
       console.warn("Behavior collapse-expand 的 trigger 参数不合法，请输入 'click' 或 'dblclick'");
     }
+    console.log('triggertrigger', trigger);
     return {
       [`node:${trigger}`]: 'onNodeClick',
     };